+++ docs/src/pages/[platform]/components/passwordfield/passwordFieldPropControls.tsx
@@ -141,7 +141,7 @@ export const PasswordFieldPropControls: PasswordFieldPropControlsInterface = ({
 
       <SwitchField
         label="hasError"
-        defaultChecked={hasError}
+        isChecked={hasError}
         labelPosition="end"
         onChange={(event) => {
           setHasError(event.target.checked as PasswordFieldProps['hasError']);
@@ -150,7 +150,7 @@ export const PasswordFieldPropControls: PasswordFieldPropControlsInterface = ({
 
       <SwitchField
         label="isDisabled"
-        defaultChecked={isDisabled}
+        isChecked={isDisabled}
         labelPosition="end"
         onChange={(event) => {
           setIsDisabled(
@@ -161,7 +161,7 @@ export const PasswordFieldPropControls: PasswordFieldPropControlsInterface = ({
 
       <SwitchField
         label="isReadOnly"
-        defaultChecked={isReadOnly}
+        isChecked={isReadOnly}
         labelPosition="end"
         onChange={(event) => {
           setIsReadOnly(
@@ -172,7 +172,7 @@ export const PasswordFieldPropControls: PasswordFieldPropControlsInterface = ({
 
       <SwitchField
         label="isRequired"
-        defaultChecked={isRequired}
+        isChecked={isRequired}
         labelPosition="end"
         onChange={(event) => {
           setIsRequired(
@@ -194,7 +194,7 @@ export const PasswordFieldPropControls: PasswordFieldPropControlsInterface = ({
 
       <SwitchField
         label="labelHidden"
-        defaultChecked={labelHidden}
+        isChecked={labelHidden}
         labelPosition="end"
         onChange={(event) => {
           setLabelHidden(
@@ -267,7 +267,7 @@ export const PasswordFieldPropControls: PasswordFieldPropControlsInterface = ({
 
       <SwitchField
         label="hideShowPassword"
-        defaultChecked={hideShowPassword}
+        isChecked={hideShowPassword}
         labelPosition="end"
         onChange={(event) => {
           setHideShowPassword(