+++ packages/react-core/src/Authenticator/context/index.ts
@@ -1,0 +1,2 @@
+export { default as AuthenticatorProvider } from './AuthenticatorProvider';
+export { AuthenticatorContext } from './AuthenticatorContext';