+++ packages/react/src/components/Authenticator/SetupTOTP/__tests__/SetupTOTP.test.tsx
@@ -1,18 +1,18 @@
 import * as React from 'react';
 import { act, render } from '@testing-library/react';
-import QRCode from 'qrcode';
-import { Auth } from 'aws-amplify';
 
-import { SetupTOTP } from '../SetupTOTP';
-import { getTotpCode } from '@aws-amplify/ui';
+import * as UI from '@aws-amplify/ui';
+import { useAuthenticator } from '@aws-amplify/ui-react-core';
+
+import { SetupTOTP } from '..';
+
+jest.mock('qrcode');
 
 jest.mock('../../hooks/useFormHandlers', () => ({
   useFormHandlers: () => ({ handleChange: jest.fn(), handleSubmit: jest.fn() }),
 }));
 
-jest.mock('../../hooks/useAuthenticator', () => ({
-  useAuthenticator: () => ({ _state: {}, isPending: false }),
-}));
+jest.mock('@aws-amplify/ui-react-core');
 
 jest.mock('../../hooks/useCustomComponents', () => ({
   useCustomComponents: () => ({
@@ -26,84 +26,65 @@ jest.mock('../../hooks/useCustomComponents', () => ({
 
 jest.mock('../../shared/FormFields', () => ({ FormFields: () => null }));
 
-const mockUser = { username: 'username' };
-const mockContext = { formFields: { setupTOTP: { QR: null } }, user: mockUser };
-jest.mock('@aws-amplify/ui', () => ({
-  ...(jest.requireActual('@aws-amplify/ui') as {}),
-  getActorState: () => ({ context: mockContext }),
-}));
-
 const DEFAULT_TOTP_ISSUER = 'AWSCognito';
 const SECRET_KEY = 'secretKey';
 
-const setupTOTPSpy = jest.spyOn(Auth, 'setupTOTP');
-const toDataURLSpy = jest.spyOn(QRCode, 'toDataURL');
+const mockUser = { username: 'username' };
+
+const getTotpCodeURLSpy = jest.spyOn(UI, 'getTotpCodeURL');
 
 describe('SetupTOTP', () => {
+  let mockGetTotpSecretCode: jest.Mock;
   beforeEach(() => {
-    jest.resetAllMocks();
+    jest.clearAllMocks();
 
-    setupTOTPSpy.mockResolvedValue(SECRET_KEY);
+    mockGetTotpSecretCode = jest.fn().mockResolvedValue(SECRET_KEY);
+
+    (useAuthenticator as jest.Mock).mockReturnValue({
+      isPending: false,
+      user: mockUser,
+      getTotpSecretCode: mockGetTotpSecretCode,
+    });
   });
 
   it('handles an undefined value when looking up its form field values', async () => {
-    const defaultTotpCode = getTotpCode(
-      DEFAULT_TOTP_ISSUER,
-      mockUser.username,
-      SECRET_KEY
-    );
-
     await act(async () => {
       render(<SetupTOTP className="className" variation="default" />);
     });
 
-    expect(setupTOTPSpy).toHaveBeenCalledTimes(1);
-    expect(setupTOTPSpy).toHaveBeenCalledWith(mockUser);
-
-    expect(toDataURLSpy).toHaveBeenCalledTimes(1);
-    expect(toDataURLSpy).toHaveBeenCalledWith(defaultTotpCode);
+    expect(getTotpCodeURLSpy).toHaveBeenCalledTimes(1);
+    expect(getTotpCodeURLSpy).toHaveBeenCalledWith(
+      DEFAULT_TOTP_ISSUER,
+      mockUser.username,
+      SECRET_KEY
+    );
   });
 
   it('handles custom values passed as form field values', async () => {
     const customTotpIssuer = 'customTOTPIssuer';
     const customTotpUsername = 'customTotpUsername';
 
-    mockContext.formFields.setupTOTP.QR = {
-      totpIssuer: customTotpIssuer,
-      totpUsername: customTotpUsername,
-    };
-
-    const customTotpCode = getTotpCode(
-      customTotpIssuer,
-      customTotpUsername,
-      SECRET_KEY
-    );
+    (useAuthenticator as jest.Mock).mockReturnValue({
+      isPending: false,
+      fields: {
+        QR: {
+          totpIssuer: customTotpIssuer,
+          totpUsername: customTotpUsername,
+        },
+      },
+      getTotpSecretCode: mockGetTotpSecretCode,
+      user: mockUser,
+    });
 
     await act(async () => {
       render(<SetupTOTP className="className" variation="default" />);
     });
 
-    expect(setupTOTPSpy).toHaveBeenCalledTimes(1);
-    expect(setupTOTPSpy).toHaveBeenCalledWith(mockUser);
-
-    expect(toDataURLSpy).toHaveBeenCalledTimes(1);
-    expect(toDataURLSpy).toHaveBeenCalledWith(customTotpCode);
-  });
-
-  describe('QR Tests', () => {
-    it('handles customTotpIssuer with spaces', async () => {
-      const customTotpIssuer = 'customTOTPIssuer spaces';
-      const customTotpUsername = 'customTotpUsername';
-
-      const customTotpCode = getTotpCode(
-        customTotpIssuer,
-        customTotpUsername,
-        SECRET_KEY
-      );
-
-      expect(customTotpCode).toBe(
-        'otpauth://totp/customTOTPIssuer%20spaces:customTotpUsername?secret=secretKey&issuer=customTOTPIssuer%20spaces'
-      );
-    });
+    expect(getTotpCodeURLSpy).toHaveBeenCalledTimes(1);
+    expect(getTotpCodeURLSpy).toHaveBeenCalledWith(
+      customTotpIssuer,
+      customTotpUsername,
+      SECRET_KEY
+    );
   });
 });